Capture rogue dependency resolution as a custom value
When dependency substitution is used, dependency configurations may be resolved
in order to determine the task graph. This can lead to confusing build scan
numbers, as well as making the build slower due to serial, rather than
parallel, resolution of configurations.

This script will report any configurations resolved this way as a custom value.