Run shellcheck on pushes AND on pull-request syncronize events
The Shellcheck Github action will _only_ comment on a PR during the
pull_request event, so let's see what happens if we trigger on both push and
pull_request. The `synchronize` action is what triggers when you push new
commits to a branch that is part of a pull request.