Restructure project to make it more friendly to end-users
End users are more interested in how to download the scripts than how the scripts are packaged together. Some users may also be interested in looking at the script source code. The new structure makes both of these things easier. The structure is less clutered.

The new structure doesn't follow the "typical" Gradle project layout, but this is also not a typical project.