Add fetch-build-validation-data-all.jar to the distribution files
Unfortunately, Gradle included builds do not yet support substituting a project
dependency that is not in the included build's "default" dependency
configuration. The shadowed jar is added to a "shadow" dependency
configuration.

So, for now, the distribution-build's assemble tasks depend on the
fetch-build-validation-data `shadowJar` task, and they use an explicit
filesystem reference to copy the generated jar in the archive file. This works,
albet it is a bit messy, and is likely to be fragile.