Updates the capture-build-scan Maven extension to capture the project name
Getting the project by listening for the MavenExecutionResult event may not be
the _best_ way to get the project, but it seems to work (at least based on some
limited testing).

The project cannot be outright injected and accessed directly because the
current MavenProject is a "scoped" object (attempting to just have it injected
fails with a Guice scope-related error).

I'm not sure if the MavenExecutionResult event can occur in parallel to the
BuildScanPublished event, so to be safe, access to the rootProjectName mutable
property is synchronized.