Fix init scripts to work when run against builds with included builds
If a Gradle build has any included builds, then the init scripts are executed
against the main build and the included builds. However, we don't want to
configure GE on those builds, as it has already been configured on the
top-level build.

But if we throw an exception if the build is not a top-level build, then the
entire experiment is halted.

Note that the init script is run against the main build AND the included
builds. So everything will still work correctly if execution on the included
builds is skipped.