Apply configuration changes after settings or root project evaluated
If the configuration changes are applied before the settings or root project
have been evaluated, then any settings applied in the build script will
override those set by the plugin, but we want the opposite when applying
configuration via system properties (we want the system property values to
override anything the user has set in their settings.gradle or build.gradle).

The callbacks for settingsEvaluated and projectEvaluated are still made within
the "withPlugin" event. This way, the configuration logic will only execute if
the Gradle Enterprise plugin has been applied.