Use the same experiment info logic for all experiments
Managing how build data is presented was becoming complex across experiments,
and error prone.

Much of the build data presented in the summary is stored in a set of bash
arrays which are declared at the top of lib/build_scan.sh. These arrays are
populated both when running a local build and from data fetched from a build
scan.

The RUN_ID is now only displayed if it is defined. Since we don't want to
define it in every experiment, it is defined in the main experiment scripts
(except for the ones where a run ID doesn't make sense).

extra_args and project_dir are NOT treated like build data because they can
never be fetched from build scans. We want to display "<not available>" for
these two properties from experiments that only fetch build data. Such
experiments (like exp4-gradle), set the value for these two settings at the top
of the script.  To make sure they are not overwritten, process_arguments has
been updated to only capture argument values if the arguments have been
defined.