Remove support for older versions of IDEA which used a "idea.versionYYYY.NN` formatted system property
For Gradle, there isn't an easy method for accessing such a property and also
remaining configuration-cache compatible. Only about 12% Gradle users who use
IDEA are using one of the older verisons of IDEA, so removing support for it
should be relatively safe.

Removed from the Maven extension as well to keep the two plugins/extensions consistent.