No longer wrap project property reads in a configuration-time provider
There is no direct configuration cache support for project-level properties
because mechanisms in which project properties are set should already cause the
configuration cache to become invalid. For example, if a project property is
set based on data in a file, or based on an environment variable, then reading
the file or the environment variable should be done in a configuraiton-cache
safe way so that if the file or the environment variable changes the
configuration cache will be invalidated.

The one exception to this is properties which are set or overriden on the
command-line. To cover this case, we still declare the property is used at
configuration time by using ProviderFactory.gradleProperty().