Fix configuration cache support by making Action/Consumer static
captureGitMetadata passes a function to BuildScanExtension.background. In order
to support configuration caching, that function (Action/Consumer) needs to be
serializable (the serialized version of the function is stored in the
configuration cache).

If the function is _not_ static, then the resulting function is actually an
anonymous inner class of CustomBuildScanEnhancements, which means the
serialized function also contains a serialized copy of the
CustomBuildScanEnhancmeents instance (thus, CustomBuildScanEnhancmenets also
needs to be serializable). However, CustomBuildScanEnhancements has a reference
to a Gradle object, which is _not_ serializable. Thus, this breaks
compatability with the configuration cache.

By making the Action/Consumer static, it is disassociated with any particular
instance of CustomBuildScanEnhancements, and therefore the lambda can be
serialized (and is therefore configuration cache compatible).

For consistency, the same modification is made to the CCUD Maven extension's
version of the class, even though Maven is not affected by this.