Generate a "fat" jar for the CCUD Maven extension
In addition to adding the extension to the extensions.xml file, the CCUD Maven
extension can be used by putting it in the MAVEN_HOME/lib/ext directory, or by
referencing the jar on the command line by setting the `maven.ext.class.path`
system property when invoking Maven.

Unfortunately, if the extension is used via one of these two additonal methods,
then Maven is unable to resolve any dependencies of the CCUD Maven extension.
That means the extension could fail with a ClassDefNotFound error.

Users could add all of the extensions dependencies to their MAVEN_HOME/lib
folder, but we can make it easier on users if we publish a jar that contains
all of the extension's dependencies (which is currently only Groovy).