Look for mvn command if the Maven wrapper is missing
It is only necessary to look for the "mvn" command by looking at the system
path. Since Maven 3.5, the M2_HOME env var was deprecated and its use is no
longer encouraged.

Using `command` to test to see if a command exists or not is more robust then
using `which` because which's behavior changes across platforms.