Remove race condition with git commit id link
We capture git metadata in a buildScan.background process, so in
general the server.url value has been set before we need it to generate
a custom link. However this is a race condition.

By not scheduling the background process until settingsEvaluated,
we avoid this race condition (at the expense of not having any Git
metadata generated if settings evaluation fails).