Redact any user:password contained in Git repo URL
This user information should not be included as part of the 'Git Repository'
custom value. With this change, any `user:password@` part of the URL will
be replaced by `******@`.

This change was based on PR #160.

Normalizing all 'Git Repository' values is not part of this fix. A later change
could potentially normalize all Git connection URLs to the canonical Git Repository URL.