Use 'buildScan.buildFinished' for deferred actions
To ensure that the correct build scan server is used when generating
custom value search links, we need to wait until no further changes will
be made to the server configuration.

Instead of `settingsEvaluated` and `afterEvalutate`, this change makes use
of `BuildScanExtension.buildFinished` to defer this work. This ensures that even
if the server URL is set during build execution the correct value will be used.

For consistency, a similar change was made to the Maven extension. This will ensure that
custom links will be generated correctly if the server URL is set in a separate custom
extension applied after the CCUD Maven extension.