gradle exp 01: save a snapshot of the build dir between the 1st build and the 2nd build
This will make it easier for users to compare build outputs between the two
builds, which users often need to do to troubleshoot why their build was not
incremental.

This also ensures that the build is executed in a folder that does not match
the git repository name (which can help identify build problems that can occur
if the dir does not match the project name).

Also moves project dir copying to a new lib (since it isn't really related to
git).