Change Kotlin CI conditional to only check for env presence
The Kotlin CI conditional is inconsistent with the Groovy sample. The
Kotlin sample checks that the environment variable "CI" exists *and* is
valued. The Groovy sample only checks for the presence of the "CI"
environment variable.

Just the presence of the "CI" environment variable should be good
enough. If a user needs it to be more specific, they can adjust the
conditional to their needs as the comment suggests.