Add quarkus-artifact.properties as an output of quarkus:build
When running ITs for native, we need quarkus-artifact.properties to be
around so it needs to be part of the cache.

Also make it an input of Failsafe.
This file contains a relative path to the binary to run and a few
additional metadata such as the full GraalVM version.
Whenever this version has changed, we need to run the tests again so it
makes sense to make it an input.

Note that we will need https://github.com/quarkusio/quarkus/pull/41587
to be in so that quarkus-artifact.properties doesn't contain a
timestamp.

Signed-off-by: Guillaume Smet <guillaume.smet@gmail.com>